+++ SQF/dayz_code/cfgVehicles.hpp
@@ -659,6 +659,29 @@ class CfgVehicles {
 		transportMaxMagazines = 100;
 		transportmaxbackpacks = 15;
 	};
+	class Ka137_PMC;
+	class Ka137_DZ: Ka137_PMC
+	{
+		displayname = "Ka137 Mini-Heli";
+		displaynameshort = "Ka137";
+		scope = 2;
+		side = 2;
+		crew = "";
+		radartype = 0;
+		typicalCargo[] = {};
+		hiddenSelections[] = {};
+		class TransportMagazines{};
+		class TransportWeapons{};
+		weapons[] = {};
+		magazines[] = {};
+		gunnerHasFlares = false;
+		commanderCanSee = 2+16+32;
+		gunnerCanSee = 2+16+32;
+		driverCanSee = 2+16+32;
+		transportMaxWeapons = 1;
+		transportMaxMagazines = 10;
+		transportmaxbackpacks = 1;
+	};
 	class House {
 		class DestructionEffects;
 	};