+++ SQF/dayz_code/rscTitles.hpp
@@ -87,7 +87,7 @@ class RscDisplayMain : RscStandardDisplay
 		class DAYZ_Version : CA_Version
 		{
 			idc = -1;
-			text = "DayZLingor 2.0 (Original Map by IceBreakr)";
+			text = "DayZLingor 2.1 (Original Map by IceBreakr)";
 			y = "(SafeZoneH + SafeZoneY) - (1 - 0.95)";
 		};
 		class CA_TitleMainMenu;